Dropship warehouse — merge AD_OrgInfo_DropShipWarehouse into AD_OrgInfo; add @CalloutMethod
The framework requires the callout class for AD_OrgInfo to be named
AD_OrgInfo (AnnotatedCalloutInstanceFactory:170-175). The dropship-warehouse
validation now lives alongside the existing AD_OrgInfo interceptor methods,
with both @ModelChange (save-time enforcement) and @CalloutMethod (immediate
UI feedback when the user selects a non-dropship warehouse in the lookup).

Resolves the Q1 code-review item from option-1 execution.
